Robot arm update

+++ Module_2/Deel_1/robotarm-python-2023-2.3/example6.py
@@ -7,26 +7,20 @@ robotArm.moveRight()
 robotArm.grab()
 robotArm.moveLeft()
 robotArm.drop()
-robotArm.moveRight()
-robotArm.grab()
-robotArm.moveRight()
-robotArm.drop()
-robotArm.moveLeft()
-robotArm.grab()
-robotArm.moveLeft()
-robotArm.drop()
-robotArm.moveRight()
-robotArm.grab()
-robotArm.moveRight()
-robotArm.drop()
-robotArm.moveLeft()
-robotArm.grab()
-robotArm.moveLeft()
-robotArm.drop()
-robotArm.moveRight()
-robotArm.grab()
-robotArm.moveRight()
-robotArm.drop()
+
+for x in range(3):
+    robotArm.moveRight()
+    robotArm.grab()
+    robotArm.moveRight()
+    robotArm.drop()
+    if x < 2:
+        for y in range(1):
+            robotArm.moveLeft()
+            robotArm.grab()
+            robotArm.moveLeft()
+            robotArm.drop()
+
+
 
 # Na jouw code wachten tot het sluiten van de window:
 robotArm.wait()
